codeql: fix 4 integer-overflow findings + filter vendored SARIF
CodeQL c-cpp alerts on our own code (vendored/generated alerts handled
separately by the SARIF filter below):

- string_repeat: len*count was a 32-bit multiply feeding allocateString
  (which takes uint64_t), so repeat(64KB-string, 65537) wrapped to a
  tiny allocation while the copy loop wrote ~4GB past it — a
  script-reachable heap overflow. Widen to uint64_t * uint64_t.
- dasRaster same-channel memcpy (3 sites): num_pixels*src_ch[*sizeof]
  multiplied in int32 before promoting to the size_t length; wraps for
  >2G-element images. Cast operands to size_t. (Only 2 were flagged;
  the u8 path has the identical shape and is fixed too.)
- volume_mixer lookahead index: attack_samples*nChannels was a uint32
  multiply inside otherwise-64-bit indexing. Cast to uint64_t.
- tutorial 14_passing_arrays.c: printf("%u", owned.size) on a uint64_t
  field is varargs UB. Use %llu + cast.

Workflow: build-mode none does not honor the init paths-ignore for C/C++
results (a tests-cpp/3rdparty/doctest alert survived two master scans
with that config), so vendored/generated findings never auto-filtered.
Add a github/filter-sarif pass between analyze (upload:false) and
upload-sarif, dropping 3rdparty, stb_*.h, sqlite, minfft, and the frozen
v1 dasFormatter parser. src/parser generated lexers are NOT excluded —
our hand-written rule bodies live there. Dropped results auto-close
their alerts on the next master scan.

+++ modules/dasAudio/src/volume_mixer.h
@@ -366,7 +366,7 @@ void ma_limiter_porcess_pcm_frames_any ( ma_limiter * limiter, float * InFames, 
     uint32_t attack_samples = limiter->attack_samples;
     for ( uint64_t i=0; i!=nFrames; ++i ) {
         for ( uint32_t j=0; j!=limiter->nChannels; ++j ) {
-            float lookahead_sample = InFames[i*limiter->nChannels + j + attack_samples*limiter->nChannels];
+            float lookahead_sample = InFames[i*limiter->nChannels + j + uint64_t(attack_samples)*limiter->nChannels];
             if (fabs(lookahead_sample) * gain[j] > threshold) {
                 float desired_gain = threshold / fabs(lookahead_sample);
                 gain[j] = gain[j] * attack_coeff + desired_gain * (1 - attack_coeff);
